refactor: adapting to the new context
