+++ app/GraphQL/Mutation/User/createUserMutation.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\GraphQL\Mutation;
+namespace App\GraphQL\Mutation\User;
 
 use Folklore\GraphQL\Support\Mutation;
 use GraphQL\Type\Definition\ResolveInfo;
@@ -10,27 +10,24 @@ use GraphQL;
 
 use App\Models\User;
 
-class UserMutation extends Mutation
+class createUserMutation extends Mutation
 {
   protected $attributes = [
-    'name' => 'user',
+    'name' => 'createUser',
     'description' => 'User mutation'
   ];
 
   public function type()
   {
     return GraphQL::type('User');
-    // return Type::listOf(Type::string());
   }
 
   public function args()
   {
     return [
-      'id' => [
-        'type' => Type::getNullableType(Type::int())
-      ],
       'name' => [
-        'type' => Type::nonNull(Type::string())
+        'type' => Type::nonNull(Type::string()),
+        'rules' => ['required']
       ],
       'email' => [
         'type' => Type::nonNull(Type::string()),
@@ -45,23 +42,15 @@ class UserMutation extends Mutation
 
   public function resolve($root, $args, $context, ResolveInfo $info)
   {
-
-    $user = User::find($args['id']);
-
-    if ($user) {
-      throw with(new MessageError('Sorry, this id is being used.'));
-      // $user->fill($args);
-      // $user->password = bcrypt($args['password']);
-      // $user->save();
-      // return $user;
-    } else {
+    try {
       $user = new User();
       $user->fill($args);
       $user->password = bcrypt($args['password']);
       $user->save();
       return $user;
+
+    } catch (\Exception $e) {
+      throw with(new MessageError($e->getMessage()));
     }
-
-
   }
 }
